Add files via upload
The file "glucose_orientation.tcl" creates a two-dimensional vector between carbon 2 and carbon 5 of each sugar molecule in the three spatial subdivisions of the nanocellulose crystal. It then calculates the orientation angle with respect to the vector (-1 1) at each time step.

The file "average_glucose_orientation.tcl" performs the same operation but averages the set of angles, obtaining a central tendency value for each spatial subdivision of the crystal.